linux_requirement.txt
aiohttp==3.7.4.post0
amqp==5.0.5
ansi==0.2.0
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.6
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
colorlog==4.7.2
cryptography==3.4.7
cycler==0.10.0
daemonize==2.5.0
databases==0.4.1
decorator==4.4.2
deepmerge==0.2.1
default==0.1.2
discord-webhook==0.11.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-ace==1.0.10
django-background-tasks==1.2.5
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-compat==1.0.15
django-crispy-forms==1.11.1
django-login-required-middleware==0.6.0
django-mathfilters==1.0.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-datatables==0.6.0
dnspython==2.0.0
dulwich==0.20.20
errbot==6.1.7
exrex==0.10.5
filelock==3.0.12
fire==0.3.1
Flask==1.1.2
future==0.18.2
futures==3.1.1
gunicorn==20.0.4
idna==2.10
idna-ssl==1.1.0
importlib-metadata==3.7.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
jsons==1.4.0
kiwisolver==1.3.1
kombu==5.0.2
lazy-object-proxy==1.4.3
loguru==0.5.3
lxml==4.6.2
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
mccabe==0.6.1
multidict==5.1.0
numpy==1.20.2
Pillow==8.2.0
postgrest==0.0.1
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.8.1
pygments-markdown-lexer==0.1.0.dev39
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.3.23
sqlparse==0.4.1
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.7.4.3
typish==1.9.1
urllib3==1.26.3
validators==0.18.2
vine==5.0.0
waitress==2.0.0
wcwidth==0.2.5
WebOb==1.8.7
WebTest==2.0.35
Werkzeug==1.0.1
wordcloud==1.8.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
